Refactor brush methods following ideas tested on circle styles
