+++ examples/klokan-basic/main.js
@@ -29,7 +29,10 @@ export function main() {
   let render = Promise.all([getStyle, getTile])
     .then( ([style, tile]) => {
       let sources = { openmaptiles: tile };
+      let t0 = performance.now();
       style.layers.forEach( layer => layer.painter(ctx, 11, sources, bboxes) );
+      let t1 = performance.now();
+      console.log("example: total render time = " + (t1 - t0).toFixed(3)  + "ms");
     });
 }
