+++ src/brush-utils.js
@@ -1,3 +1,5 @@
+import { geomToPath } from "./path.js";
+
 export function canv(property) {
   // Create a default state setter for a Canvas 2D renderer
   return (val, ctx) => { ctx[property] = val; };
@@ -6,6 +8,28 @@ export function canv(property) {
 export function pair(getStyle, setState) {
   // Return a style value getter and a renderer state setter as a paired object
   return { getStyle, setState };
+}
+
+export function initBrush({ setters, methods }) {
+  const dataFuncs = setters.filter(s => s.getStyle.type === "property");
+  const zoomFuncs = setters.filter(s => s.getStyle.type !== "property");
+
+  return function(ctx, zoom, data) {
+    // Set the non-data-dependent context state
+    zoomFuncs.forEach(f => f.setState(f.getStyle(zoom), ctx));
+
+    // Loop over features and draw
+    data.features.forEach(feature => drawFeature(ctx, zoom, feature));
+  }
+
+  function drawFeature(ctx, zoom, feature) {
+    // Set data-dependent context state
+    dataFuncs.forEach(f => f.setState(f.getStyle(zoom, feature), ctx));
+
+    // Construct Path and draw
+    let path = geomToPath(feature.geometry); // TODO: move to data prep
+    methods.forEach(method => ctx[method](path));
+  }
 }
 
 export function makePatternSetter(sprite) {
